+++ packages/at_lookup/lib/src/at_lookup.dart
@@ -28,11 +28,22 @@ abstract class AtLookUp {
 
   Future<String?> executeCommand(String command, {bool auth = false});
 
-  /// performs a PKAM authentication using private key on the client side and public key on secondary server
+  /// Performs a PKAM authentication using private key on the client side and public key on secondary server
+  ///
   /// Pkam private key should be set in  [atChops.atChopsKeys]
+  ///
   /// Default signing algorithm for pkam signature is [SigningAlgoType.rsa2048] and default hashing algorithm is [HashingAlgoType.sha256]
+  ///
   /// Optionally pass enrollmentId if the client is enrolled using APKAM
   Future<bool> pkamAuthenticate({String? enrollmentId});
+
+  /// Generates digest using from verb response and [secret] and performs a
+  /// CRAM authentication to secondary server
+  Future<bool> cramAuthenticate(String secret);
+
+  /// Terminates the underlying connection to the atServer
+  /// used by this instance of AtLookup
+  Future<void> close();
 
   /// set an instance of  [AtChops] for signing and verification operations.
   set atChops(AtChops? atChops);